requirements.txt
google-genai~=1.13.0